feat: create a newtype to return the cal already wrapped
