+++ embedded-cal-nrf54l15/tests/integration.rs
@@ -4,29 +4,20 @@
 use defmt_rtt as _;
 use panic_probe as _;
 
-struct ImplementSha256Short;
-impl embedded_cal_software::ExtenderConfig for ImplementSha256Short {
-    const IMPLEMENT_SHA2SHORT: bool = true;
-    type Base = embedded_cal_nrf54l15::Nrf54l15Cal;
-}
 struct TestState {
-    cal: embedded_cal_software::Extender<ImplementSha256Short>,
+    cal: embedded_cal_nrf54l15::Nrf54l15Cal,
 }
 
 #[defmt_test::tests]
 mod tests {
-    use super::ImplementSha256Short;
     use embedded_cal_nrf54l15::Nrf54l15Cal;
 
     #[init]
     fn init() -> super::TestState {
         // FIXME: How to make sure there is a exclusive reference for CRACEN_S?
-        let base =
-            embedded_cal_nrf54l15::Nrf54l15Cal::new(nrf_pac::CRACEN_S, nrf_pac::CRACENCORE_S);
-
-        let cal = embedded_cal_software::Extender::<ImplementSha256Short>::new(base);
-
-        super::TestState { cal }
+        super::TestState {
+            cal: Nrf54l15Cal::new(nrf_pac::CRACEN_S, nrf_pac::CRACENCORE_S),
+        }
     }
 
     #[test]
@@ -40,7 +31,7 @@ mod tests {
     #[test]
     fn test_hmac_sha256(state: &mut super::TestState) {
         embedded_cal::test_hmac_algorithm_hmacsha256::<
-            <embedded_cal_software::Extender<ImplementSha256Short> as embedded_cal::HmacProvider>::Algorithm,
+            <Nrf54l15Cal as embedded_cal::HmacProvider>::Algorithm,
         >();
         testvectors::test_hmac_sha256(&mut state.cal);
     }
