+++ embedded-cal-stm32wba55/Cargo.toml
@@ -20,6 +20,7 @@ harness = false
 
 [dependencies]
 embedded-cal.path = "../embedded-cal"
+embedded-cal-software.path = "../embedded-cal-software"
 stm32-metapac = { version = "18.0", features = ["stm32wba55cg", "rt"] }
 
 [dev-dependencies]
@@ -28,6 +29,5 @@ cortex-m-rt = { version = "0.7", features = ["device"] }
 defmt = "1.0"
 defmt-rtt = "1.0"
 defmt-test = "0.4"
-embedded-cal-software.path = "../embedded-cal-software"
 panic-probe = { version = "1.0", features = ["print-defmt"] }
 testvectors.path = "../testvectors"
