+++ embedded-cal-stm32wba55/tests/integration.rs
@@ -4,34 +4,19 @@
 use defmt_rtt as _;
 use panic_probe as _;
 
-struct ImplementSha256Short;
-impl embedded_cal_software::ExtenderConfig for ImplementSha256Short {
-    const IMPLEMENT_SHA2SHORT: bool = true;
-    type Base = embedded_cal_stm32wba55::Stm32wba55Cal;
-}
-
 struct TestState {
-    /// Software-extended CAL used for the SHA-256 hash tests.
-    cal: embedded_cal_software::Extender<ImplementSha256Short>,
-    /// Raw STM32WBA55 CAL used to exercise the hardware HMAC accelerator directly.
-    raw: embedded_cal_stm32wba55::Stm32wba55Cal,
+    cal: embedded_cal_stm32wba55::Stm32wba55Cal,
 }
 
 #[defmt_test::tests]
 mod tests {
-    use super::ImplementSha256Short;
     use embedded_cal_stm32wba55::Stm32wba55Cal;
+
     #[init]
     fn init() -> super::TestState {
-        // stm32_metapac::HASH is a `const`, so it can be used at two sites
-        // without conflict; both values alias the same hardware register block.
-        // Tests run sequentially so there is no concurrent access.
-        let raw =
-            embedded_cal_stm32wba55::Stm32wba55Cal::new(stm32_metapac::HASH, &stm32_metapac::RCC);
-        let base =
-            embedded_cal_stm32wba55::Stm32wba55Cal::new(stm32_metapac::HASH, &stm32_metapac::RCC);
-        let cal = embedded_cal_software::Extender::<ImplementSha256Short>::new(base);
-        super::TestState { cal, raw }
+        super::TestState {
+            cal: Stm32wba55Cal::new(stm32_metapac::HASH, &stm32_metapac::RCC),
+        }
     }
 
     #[test]
@@ -47,7 +32,6 @@ mod tests {
         embedded_cal::test_hmac_algorithm_hmacsha256::<
             <Stm32wba55Cal as embedded_cal::HmacProvider>::Algorithm,
         >();
-        // Runs directly against the hardware HMAC accelerator (MODE=1 in HASH_CR).
-        testvectors::test_hmac_sha256(&mut state.raw);
+        testvectors::test_hmac_sha256(&mut state.cal);
     }
 }
